CAY-1666 Fix problem with tests on ingres db
git-svn-id: https://svn.apache.org/repos/asf/cayenne/main/trunk@1244438 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/cayenne-jdk1.5-unpublished/src/test/java/org/apache/cayenne/access/DataContextClobTest.java
@@ -49,7 +49,9 @@ public class DataContextClobTest extends ServerCase {
 
     @Override
     protected void setUpAfterInjection() throws Exception {
-        dbHelper.deleteAll("CLOB_TEST");
+        if (accessStackAdapter.supportsLobs()) {
+            dbHelper.deleteAll("CLOB_TEST");
+        }
     }
 
     private boolean skipTests() {
